feat: Enhance device and session activity management with new features
- Updated DeviceController to handle prices directly in Egyptian Pounds (EGP) without conversion.
- Modified SessionActivityController to validate session ID for activity retrieval and updates.
- Introduced ActivityPause and ActivityModeChange models to track pauses and mode changes in activities.
- Enhanced SessionActivityService to manage activity status changes, including pause and resume functionality.
- Updated session management to ensure device availability and handle session end logic more effectively.
- Added CreateGuestUserRequest and createGuest method in UserController for guest user creation during session initiation.
- Adjusted migrations and seeders to reflect changes in pricing and new activity tracking features.

+++ database/migrations/2025_10_21_011824_create_devices_table.php
@@ -17,8 +17,8 @@ return new class extends Migration
             $table->text('description')->nullable();
             $table->enum('device_type', ['ps4', 'ps5', 'billiard']);
             $table->enum('status', ['available', 'in_use', 'maintenance'])->default('available');
-            $table->integer('price_per_hour');
-            $table->integer('multi_price')->nullable();
+            $table->decimal('price_per_hour', 8, 2);
+            $table->decimal('multi_price', 8, 2)->nullable();
             $table->text('notes')->nullable();
             $table->timestamps();
             $table->softDeletes();
